services/order-service/requirements.txt
Django
djangorestframework
mysqlclient
requests
PyJWT
pika